+++ tests/test_cell_readers.py
@@ -894,9 +894,9 @@ def test_save_cellpyfile_auto_extension(cellpy_data_instance, parameters):
     cellpy_data_instance.make_step_table()
     tmp_file = next(tempfile._get_candidate_names())
     cellpy_data_instance.save(tmp_file)
-    assert os.path.isfile(tmp_file + ".h5")
-    os.remove(tmp_file + ".h5")
-    assert not os.path.isfile(tmp_file + ".h5")
+    assert os.path.isfile(tmp_file + ".cellpy")
+    os.remove(tmp_file + ".cellpy")
+    assert not os.path.isfile(tmp_file + ".cellpy")
 
 
 def test_save_cellpyfile_auto_extension_pathlib(cellpy_data_instance, parameters):
@@ -905,7 +905,7 @@ def test_save_cellpyfile_auto_extension_pathlib(cellpy_data_instance, parameters
     cellpy_data_instance.make_step_table()
     tmp_file = pathlib.Path(next(tempfile._get_candidate_names()))
     cellpy_data_instance.save(tmp_file)
-    tmp_file = tmp_file.with_suffix(".h5")
+    tmp_file = tmp_file.with_suffix(".cellpy")
     assert tmp_file.is_file()
     os.remove(tmp_file)
     assert not tmp_file.is_file()